requirements.txt
requests
python-mpd2
mutagen
